chore(ByFrameRepeater): extract sorting to a service
